fix: move isMissingCoords to dependency-free module to fix Vercel build
kennel-utils.ts imports prisma/pg which can't be bundled for the browser.
Created kennel-helpers.ts (no server deps) for pure functions shared
across server and client components.
