+++ src/lib/kennel-helpers.ts
@@ -1,0 +1,3 @@
+/** Check whether a kennel is missing geocoordinates. */
+export const isMissingCoords = (k: { latitude: number | null; longitude: number | null }) =>
+  k.latitude == null || k.longitude == null;
